ICI_simulateur

Documentation goes here.

Example blocks

Script: @example

A = rand(3, 3)
b = [1, 2, 3]
A \ b

3-element Vector{Float64}:
 11.197187370885707
 -0.321219844334125
 -1.7232280923537142

REPL: @repl

julia> 1 + 12

Documentation tests

Script

a = 1
b = 2
a + b

# output

3

REPL

julia> a = 1
1

julia> b = 2;

julia> c = 3;  # comment

julia> a + b + c
6

Reference

@autodocs

ICI_simulateur.Principal — Method

Function Principal Fonction principale du simulateur. Permet de lancer les simulations de propagation d'épidémies sur une région donnée. C'est cette fonction qu'il faut appeler dans le programme principal

ICI_simulateur.choix_resto — Method

Function choix_resto

ICI_simulateur.choix_step — Method

Function JourneeIndiv_trace

ICI_simulateur.choix_travail — Method

Function ChoixTravail #Fonction permettant d'associer une entreprise à un individu ainsi que le tissu social (~collègues proches) #Entrées : #- statut::Array{Int32,1} #- indivinout::Array{Int32,1} #- idwork::Array{Int32,1} #- areawork::Array{Float64,1} #Sorties : #- idindwork::Array{Int32,1} #- tissu_work::Array{Int32,1}

ICI_simulateur.compute_indicateur! — Method

function Compute_Indicateur #Fonction permettant de calculer les indicateurs épidémiologiques

ICI_simulateur.confidence_interval — Method

Confidence_interval(resultat,alpha)

# Function used to compute the confidence interval of each parameter
# of the modeling
# Input:
# - resultat:
# - alpha: alpha-confidence
# Output:
# - Interv: Matri containing the confidence interval of the indicator

ICI_simulateur.contamination_direct — Method

ContaminationDirect(depc,deps,arrc,arrs,proba)

# Function used to compute the contamination
# Input:
# - depc:
# - deps:
# - arrc:
# - arrs:
# - proba:
# Output:
# - p

ICI_simulateur.contamination_externe! — Method

Function Contaminationexterne! #Fonction permettant de calculer la contamination ayant lieu à l'extérieur du domaine #Entrées : #- statcovid::Statutcovid, #- incidence::Array{Float64,1} #- statut::Array{Int32,1} #- indivinout::Array{Int32,1} #- whereis_local::Array{Any,1} #Sorties :

ICI_simulateur.contamination_fct! — Method

Fonction Contamination_fct! #Fonction calculant la contamination au sein du quartier

ICI_simulateur.covid_identite — Method

Covid_Identite()

# Function used to define the parameters associated to the Covid for each individual
# Input:
# NO INPUT
# Output:
# - tps_incub ({Int32}): incubation time (expressed in day)
# - tps_dt_cont ({Int32}):
# - proba ({Float64}): probability of contamination (between 0 and 1)

ICI_simulateur.def_indiv_in — Method

Function IndivIn #Fonction permettant de mettre en place les individus qui entrent pour travailler au sein du quartier #Entrées : #- indivout::Array{Int32,1} #- statut::Array{Int32,1} #- idhouse::Array{Int32,1} #Sorties : #- indivinout::Array{Int32,1} #- statut::Array{Int32,1} #- idhouse::Array{Int32,1}

ICI_simulateur.def_indiv_out — Method

Function IndivOut #Fonction permettant de mettre en place les individus qui sortent pour travailler en dehors du quartier #Entrée : #- statut::Array{Int32,1} #Sorties : #- indivout::Array{Int32,1}

ICI_simulateur.def_parametre_input — Method

DefParametreInput(namefolder,namefile_graph)

Function used to define a file with all of the input data used in the modeling.
This function creates a txt file named Input_Data.txt

Input:
- name_folder::String (name of the folder where the file is created)
- name_file_graph::String (name of the graph used in the modeling)
Output:
- NaN

ICI_simulateur.def_statut — Method

Function DefStatut #Fonction permettant de définir les statuts sociales des individus : actif, étudiant, écolier, retraité #Entrées : #- cat_age::Array{Int32,1} #Sorties : #- statut::Array{Int32,1}

ICI_simulateur.deplacement_v2! — Method

Function Deplacement_v2!

Fonction permettant de définir les déplacements des individus sur toute une journée. Cette journée est différente selon le statut des individus. Si une mesure sanitaire est mise en place, alors cette journée est impactée.

Entrées :

Jour : Structure DefJournee à mettre à jour

Ville : Structure City contenant les informations sur les lieux et bâtiments

Individu : Structure Population contenant les informations sur les individus

statcovid : Structure Statut_covid contenant les informations sur la covid

Sortie :

Jour : Structure Jour contenant les lieux visités par les individus

ICI_simulateur.distance_resto — Method

Function DistanceResto Fonction permettant de définir la distance la plus courte entre une entreprise et l'ensemble des services de restaurations.

ICI_simulateur.distribution_workingplace — Method

Function RepartitionEntreprise #Fonction permettant de repartir les individus au sein des entreprises #Entrées : #- actif::Array{Int64,1} #- idwork::Array{Int32,1} #- areawork::Array{Float64,1} #Sorties : #- Repartition::Array{Int64,1}

ICI_simulateur.heure_indiv — Method

Function heure_indiv

ICI_simulateur.indiv_inout — Method

Function IndivInOut #Fonction permettant de mettre en place les individus entrant et sortant du quariter #Entrées : #- statut::Array{Int32,1} #- idhouse::Array{Int32,1} #Sorties : #- indivinout::Array{Int32,1} #- statut::Array{Int32,1} #- id_house::Array{Int32,1}

ICI_simulateur.individual — Method

Individual(graph, n_node)

# Function used to find the id node of all individuals
# Input:
# - graph: undirected Int64 metagraph with Float64 weights
# - n_node ({Int32}): total number of nodes
# Output:
# - id_indiv (Array{Int32,.}): Vector containing the id of all individuals
# - id_house (Array{Int32,.}): Vector containing the id of the corresponding house
# - cat_age (Array{Int32,.}): Vector containing the age category

ICI_simulateur.journee_indiv — Method

Function journee_indiv Fonction permettant de définir la journée d'un individu.

Entrées :

i_indiv : Indice de l'individu

Ville : Structure City contenant les informations sur les lieux et bâtiments

Individu : Structure Population contenant les informations sur les individus

ICI_simulateur.mise_a_jour! — Method

miseajour!(contamination0,contamination1,p_covid,cpt)

# Function used to update the contamination at the end of the day
# Input:
# - contamination0:
# - contamination1:
# - p_covid:
# - cpt:
# Output :
# - NO OUTPUT

ICI_simulateur.mise_en_place! — Method

Function MiseEnPlace! #Fonction permettant de mettre en place des mesures sanitaires #Entrées : #- statcovid::Statut_covid #- indicateur::Int32 #Sorties :

ICI_simulateur.miseajour_mesure! — Method

Function miseajour_mesure!

ICI_simulateur.miseajour_tracage! — Method

Function miseajour_tracage!

ICI_simulateur.parallelisation — Method

Parallelisation()

Function used to define the parallelization. The function returns the rank of the processus

Input:
- NaN
Output:
- comm
- rank::Int

ICI_simulateur.poi — Method

POI(graph, n_node)

Function used to find the id node of all POI
Input:
- graph: undirected Int64 metagraph with Float64 weights
- n_node ({Int32}): total number of nodes
Output:
- id_poi (Array{Int32,.}): Vector containing the id of all POI

ICI_simulateur.proba_restaurant — Method

Function proba_restaurant() Fonction permettant de définir une probabilité liée à la distance entre le restaurant et l'entreprise

ICI_simulateur.save_indicators — Method

Saveindicators(resultat,namefile)

# Function used to write the result of the modeling into a txt file
# Input:
# - resultat:
# - name_file: name of the file
# Output :
# NO OUTPUT

# This function needs the library DelimitedFiles

ICI_simulateur.save_indicators_jsonfile — Method

Saveindicatorsjson(resultat,name_file)

# Function used to write the result of the modeling into a txt file
# Input:
# - resultat:
# - name_file: name of the file
# Output :
# NO OUTPUT

# This function needs the library JSON

ICI_simulateur.tissu — Method

Function Tissu

ICI_simulateur.tissu_social — Method

Function Tissu_social #Fonction permettant de mettre en place un tissu social au sein de l'entreprise. Ce sont des personnes qui seront plus proches entre eux. #Entrées : #- Structure::Array{Int32,1} #Sorties : #- Structure::Array{Int32,1}

ICI_simulateur.working_place — Method

WorkingPlace(graph, n_node)

Function used to find the id node of all working place
Input:
- graph: undirected Int64 metagraph with Float64 weights
- n_node ({Int32}): total number of nodes
Output:
- id_work (Array{Int32,.}): Vector containing the id of all working place
- area_work (Array{Int32,.}): Vector containing the area of each working place

ICI_simulateur.DefJournee — Type

Struct DefJournee #Structure contenant les informations des déplacements des individus (identifiant du lieu, heure d'arrivée et heure de départ)

ICI_simulateur.Simulateur — Type

Struct Simulateur

ICI_simulateur.Statut_covid — Type

mutable struct Statut_Covid #Structure contenant les paramètres épidémiologiques du modèle